Typos in chapter 6
This book is really nice! 
Here are some mistakes or incosistencies I noticed while reading.

I'm not shure I understand what this sentence means: "3. We register an interest in `Read` events on this socket with a second syscall to register our interest in the event queue we created in (1)." 
Could you rephrase it?